+++ modules/apps/blogs/blogs-uad-test/src/testIntegration/java/com/liferay/blogs/uad/test/util/BlogsEntryUADTestUtil.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.blogs.uad.test;
+package com.liferay.blogs.uad.test.util;
 
 import com.liferay.blogs.model.BlogsEntry;
 import com.liferay.blogs.service.BlogsEntryLocalService;
